proof-read README.md
There were some English mistakes in README.md, I did a proof-reading pass to fix them.

I see other issues with the README text as follows. In particular the text in the beginning of the "Usage" subsection is very hard to understand and out of date (it mentions a `parse_decorated` "module" that seems to have been removed from Cairn)

It looks like the "Detailed usage example" has the same information, but with code instead of unclear words, maybe you could remove the informal text at the beginning '"Usage" and have only the detailed usage example?

Ideally I think it would be nice to be a bit more explicit: start from a very simple repo with a very simple Menhir grammar -- show the file layout to users -- and then show the commands and code needed add Cairn support to this repo (explaining roughly what they do), and then show a textual capture of what happens when you run it. Ideally, this minimal setup would also be included in an `examples/` subdirectory somewhere, so that people can also clone it and then play with it.